requirements.txt
asgiref==3.4.1
Django==3.2.11
django-rest-framework==0.1.0
djangorestframework==3.13.1
pytz==2021.3
sqlparse==0.4.2
typing_extensions==4.0.1
djangorestframework-simplejwt==5.2.2
psycopg2
minio
